refactor: classify GraspGenX flow as demo
