+++ .github/workflows/ci.yml
@@ -306,7 +306,7 @@ jobs:
         run: |
           git config --local http.https://github.com/.extraheader \
             "AUTHORIZATION: basic $(printf 'x-access-token:%s' '${{ github.token }}' | base64 -w0)"
-          git lfs pull --include="dimos/manipulation/graspgenx_demo/fixtures/ycb_banana_scene.npz"
+          git lfs pull --include="dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.npz"
           git config --local --unset-all http.https://github.com/.extraheader
       - name: Install uv
         uses: astral-sh/setup-uv@v8.3.2
@@ -536,7 +536,7 @@ jobs:
         run: |
           git config --local http.https://github.com/.extraheader \
             "AUTHORIZATION: basic $(printf 'x-access-token:%s' '${{ github.token }}' | base64 -w0)"
-          git lfs pull --include="dimos/manipulation/graspgenx_demo/fixtures/ycb_banana_scene.npz"
+          git lfs pull --include="dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.npz"
           git config --local --unset-all http.https://github.com/.extraheader
       - name: Fix permissions
         run: |
@@ -709,7 +709,7 @@ jobs:
         run: |
           git config --local http.https://github.com/.extraheader \
             "AUTHORIZATION: basic $(printf 'x-access-token:%s' '${{ github.token }}' | base64 -w0)"
-          git lfs pull --include="dimos/manipulation/graspgenx_demo/fixtures/ycb_banana_scene.npz"
+          git lfs pull --include="dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.npz"
           git config --local --unset-all http.https://github.com/.extraheader
       - name: Fix permissions
         run: |
