+++ .github/workflows/release-build-check.yml
@@ -12,7 +12,7 @@ on:
       - setup.py
       - MANIFEST.in
       - .gitattributes
-      - dimos/manipulation/graspgenx_demo/**
+      - dimos/manipulation/demo_graspgenx/**
 
 permissions: {}
 
@@ -29,7 +29,7 @@ jobs:
       - name: Checkout
         uses: actions/checkout@v7
       - name: Hydrate GraspGenX fixture
-        run: git lfs pull --include="dimos/manipulation/graspgenx_demo/fixtures/ycb_banana_scene.npz"
+        run: git lfs pull --include="dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.npz"
       - name: Build one wheel
         uses: pypa/cibuildwheel@v4.1.0
         with:
