+++ dimos/manipulation/demo_graspgenx/extractor.py
