+++ dimos/manipulation/demo_graspgenx/fixture.py
@@ -104,7 +104,7 @@ def _load_arrays(path: Path) -> tuple[np.ndarray, np.ndarray, float]:
     if path.read_bytes()[: len(LFS_POINTER_PREFIX)] == LFS_POINTER_PREFIX:
         raise ValueError(
             f"scene fixture {path} is a Git-LFS pointer, not hydrated NPZ data; "
-            'run `git lfs pull --include="dimos/manipulation/graspgenx_demo/fixtures/ycb_banana_scene.npz"`'
+            'run `git lfs pull --include="dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.npz"`'
         )
     with np.load(path, allow_pickle=False) as data:
         points = np.asarray(data["points"], dtype=np.float32)
