+++ dimos/manipulation/demo_graspgenx/fixtures/ycb_banana_scene.json
