+++ openspec/specs/graspgenx-inference-demo/spec.md
@@ -17,7 +17,7 @@ The demo SHALL use a deterministic recorded YCB banana scene fixture containing 
 
 ### Requirement: Report and validate real inference
 
-The direct contributor tool MUST report checkpoint, CUDA/device, scene point count, object point count, candidate count, best score, result frame, and output path. It MUST validate nonempty finite ranked TCP-frame output with the expected header and successful YAML output. The adapter contract and its nonidentity unit test verify TCP calibration by right-multiplication; the contributor layer does not independently observe raw backend transforms. Contributors invoke it with `uv run --extra graspgenx python -m dimos.manipulation.graspgenx_demo`. The tool MUST run fixture loading, the real adapter, ranked candidate generation, and atomic YAML/RRD publication in one direct process; it MAY launch a detached viewer. Its exit status MUST be zero only after successful validation and publication, and nonzero for configuration, inference, validation, or output failures. This requirement does not provide evidence of worker, RPC, coordinator, or blueprint deployment coverage.
+The direct contributor tool MUST report checkpoint, CUDA/device, scene point count, object point count, candidate count, best score, result frame, and output path. It MUST validate nonempty finite ranked TCP-frame output with the expected header and successful YAML output. The adapter contract and its nonidentity unit test verify TCP calibration by right-multiplication; the contributor layer does not independently observe raw backend transforms. Contributors invoke it with `uv run --extra graspgenx python -m dimos.manipulation.demo_graspgenx`. The tool MUST run fixture loading, the real adapter, ranked candidate generation, and atomic YAML/RRD publication in one direct process; it MAY launch a detached viewer. Its exit status MUST be zero only after successful validation and publication, and nonzero for configuration, inference, validation, or output failures. This requirement does not provide evidence of worker, RPC, coordinator, or blueprint deployment coverage.
 
 #### Scenario: CUDA smoke run succeeds
 - **GIVEN** a compatible checkpoint, fixed sweep-volume gripper configuration, and available CUDA device
